SC Qt5 upgraded to Qt6.9

+++ src/imageproc/GrayImage.h
@@ -58,7 +58,7 @@ class GrayImage {
    * An additional guaranee provided by the underlying QImage
    * is that this value is a multiple of 4.
    */
-  [[nodiscard]] int stride() const { return m_image.bytesPerLine(); }
+  [[nodiscard]] size_t stride() const { return m_image.bytesPerLine(); }
 
   [[nodiscard]] QSize size() const { return m_image.size(); }
